Update README.md
The purpose of this assignment is to introduce you to dynamic memory allocation in C. You will also learn how to implement a singly-linked list data structure in C.

How do we allocate memory on the heap? How do we de-allocate it when it is no longer used?

The goal of this assignment is to learn how to dynamically allocate and de-allocate memory, or manage memory, in C.

1.2 Task

In this assignment, you will be implementing a singly-linked list data structure. Your linked list nodes will have struct data (of type User*). Each User struct contains a union with data for either a Student or an Instructor. You can find details about this struct and the contained union in the included list.h file. Your linked list will be able to add, remove, and mutate, and query the data stored within it.

You will be writing code in

list.c